Move usaspending task to tasks.py
